requirements.txt
Flask==3.0.2
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.27
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
lxml==4.9.3
pandas==2.0.3
chromadb==0.4.24
openai==1.51.0
